Add Intermediate Stage to Menu Hide, Alter Button
Adds an intermediate stage to the Menu Hide feature, removing only the
non-functional aspects. This allows to see the kobold better AND keep
editing them without having to unhide and rehide the menu repeatedly.

Altered button to fit new functionality, making it match the color
scheme of the other buttons and instead displaying a "indicator light"
with the current stage of hiding next to it instead. Button is now
smaller to fit, does unfortunately fail to line up as well text-wise.

Also, moved the reset all kobold settings button to the bottom. It just
feels better there personally. In future may also add additional spacing
between the button and the last option somehow, and enlarge the title of
the menu (and maybe change it from "Kobold" to "Default Kobold" or
"Kobold Options")