Make mozzarella linerenderer use world length UVs
Per the docs: "Repeat the texture along the line, based on its length in world units."
Continuing, for now, to make changes directly in the project's local copy of this package, because I don't know what's up with the "real" one?
Also, just hardcoding it to always use this mode, since I believe it's only used for this one purpose (shooting fluids) everywhere in the project.  Could put it behind a public boolean or subclass if necessary later though.  (But the same hardcoded change was already approved for generateLightingData so I assume this makes sense.)