Remove color from matcap emission
One more change while I was in there -- I think mixing the color into the emission is a mistake?  It was causing all fluids to almost totally lose their shading, and highlights to get completely washed out; I've replaced it with just passing the matcap directly into the emission value, so only the parts actually shining in the matcap are affected, rather than even the darkest parts always *also* doubling whatever the input color value was supposed to be.   However, if this was a deliberate aesthetic choice for some reason, I can undo just this last commit.