Fixes the issue that happens when you have many equipment mods
When you have a lot of equipment mods, if you use the /equip command you get the wrong equipment, this can also affect mods equipments.

With this fix that no longer happens but if you have a mod with duplicate equipment or equipment with the same name as other equipment you will not be able to load that or those mods in the mod list.