Expanded GameEventResponses Again
AnimatorMathInteger - Similar to AnimatorSetInteger, but Adds the specified Value to the currentValue of the specified Integer. Value can be negative, instead Subtracting from the currentValue.
ConditionalCountEnabled - Counts the number of enabled trackedObjects, then invokes two different Response sets, dependent on whether or not the total enabledCount is equal to the Requirement. Can be inverted to instead track disabled objects, instead passing the Requirement if enabledCount is lesser than or equal to the Requirement.